requirements.txt
opencv-python==4.8.1.78
speedtest-cli==2.1.3
aiohttp==3.8.6
python-telegram-bot==20.6
python-dotenv==1.0.0
schedule==1.2.1
fastcli==0.1.1
openmeteo-requests==1.1.0
requests-cache==1.1.0
retry-requests==2.0.0
ping3==4.0.4
APScheduler==3.10.4
nest-asyncio==1.5.8
beautifulsoup4==4.12.2
lxml==4.9.3



